# docs/swagger.yaml
definitions:
...
  e_time_template.TimeDatum:
    properties:
      condition_type:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      interval_seconds:
        type: integer
      repeat_type:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      t_condition:
        items:
          type: integer
        type: array
    type: object
  e_time_template.TimeDatumCreate:
    properties:
      condition_type:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      interval_seconds:
        type: integer
      repeat_type:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      t_condition:
        items:
          type: integer
        type: array
    required:
    - end_time
    - start_date
    - start_time
    - t_condition
    type: object
  e_time_template.TimeDatumUpdate:
    properties:
      condition_type:
        type: string
      end_date:
        type: string
      end_time:
        type: string
      interval_seconds:
        type: integer
      repeat_type:
        type: string
      start_date:
        type: string
      start_time:
        type: string
      t_condition:
        items:
          type: integer
        type: array
    required:
    - end_time
    - repeat_type
    - start_date
    - start_time
    - t_condition
    type: object
  e_time_template.TimeTemplate:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      time_data:
        $ref: '#/definitions/e_time_template.TimeDatum'
      updated_at:
        type: string
    type: object
  e_time_template.TimeTemplateCreate:
    properties:
      name:
        type: string
      time_data:
        $ref: '#/definitions/e_time_template.TimeDatumCreate'
    required:
    - name
    - time_data
    type: object
  e_time_template.TimeTemplateUpdate:
    properties:
      id:
        type: integer
      name:
        type: string
      time_data:
        $ref: '#/definitions/e_time_template.TimeDatumUpdate'
    required:
    - id
    type: object
...
host: 127.0.0.1:5487
info:
  contact:
    email: wwilson008@gmail.com
    name: Wilson
    url: https://github.com/littlebluewhite
  description: This is a schedule-command server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Schedule-Command swagger API
  version: "2.0"
paths:
...
  /time_template/:
    delete:
      parameters:
      - description: time_template id
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: delete successfully
          schema:
            type: string
      summary: Delete time templates
      tags:
      - time_template
    get:
      description: Get all time templates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/e_time_template.TimeTemplate'
            type: array
      summary: Show all time templates
      tags:
      - time_template
    patch:
      consumes:
      - application/json
      parameters:
      - description: modify time template body
        in: body
        name: time_template
        required: true
        schema:
          items:
            $ref: '#/definitions/e_time_template.TimeTemplateUpdate'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: update successfully
          schema:
            type: string
      summary: Update time templates
      tags:
      - time_template
    post:
      consumes:
      - application/json
      parameters:
      - description: time template body
        in: body
        name: time_template
        required: true
        schema:
          items:
            $ref: '#/definitions/e_time_template.TimeTemplateCreate'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/e_time_template.TimeTemplate'
            type: array
      summary: Create time templates
      tags:
      - time_template
  /time_template/{id}:
    get:
      description: Get time templates by id
      parameters:
      - description: time template id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/e_time_template.TimeTemplate'
      summary: Show time templates
      tags:
      - time_template
swagger: "2.0"